getBudgetById + correct portefeuilleimpl cache

+++ app/src/main/java/fr/budgethashtag/service/portefeuille/PortefeuilleService.java
@@ -6,5 +6,5 @@ public interface PortefeuilleService extends MotherService {
     void getOrCreateDefaultPortefeuilleIfNotExistAsync();
     void getPortefeuilleByIdAsync();
 
-    long getIdPortefeuilleFromSharedPref();
+    int getIdPortefeuilleFromSharedPref();
 }
